Analyzing stocks trend to determine market volatility
Analyzing the prices of major stock data and fit an ARIMA model to make predictions of the stock prices.  The stocks used in this analysis are  Visa, MasterCard,  Bank of America,  America Airline, and Wells Fargo stocks. The ARIMA model is fitted for the Bank of America dataset.
The data
The data used in the analysis is derived from Yahoo finance . To get the data we use python which has inbuilt packages that can be used to collect the stock prices.  Only the stock ticker is obtained and used to analyze the data from yahoo.  The dates for the collection start from 2010 January 1st to 24th march 2023.  The data is then stored into an excel file then used for analysis.